# backend/swagger.yaml
swagger: "2.0"
info:
  title: KnightWise
  description: |
    API documentation for KnightWise, the computer science guided learning platform.

    ## Question Types and Grading
    
    KnightWise supports five non-programming question types, each with its own grading algorithm:
    
    ### Multiple Choice
    - **Input Format**: String (e.g., `"Paris"`)
    - **Grading**: Exact match (case-insensitive, trimmed)
    - **Normalized Score**: 1.0 (correct) or 0.0 (incorrect)
    
    ### Fill-In-the-Blanks
    - **Input Format**: String (e.g., `"Paris"`)
    - **Grading**: Exact or fuzzy matching using Levenshtein distance
    - **Normalized Score**: 0.0 to 1.0 based on similarity (1 - distance/maxLength)
    - **Partial Credit**: Yes (>= 50% similarity, API feedback says "Almost correct!")
    
    ### Select-All-That-Apply
    - **Input Format**: Array of strings (e.g., `["Python", "JavaScript"]`)
    - **Grading**: (correct selections - incorrect selections) / total correct options
    - **Normalized Score**: 0.0 to 1.0 based on how many correct selections
    - **Partial Credit**: Yes
    
    ### Ranked Choice
    - **Input Format**: Array of strings in ranked order (e.g., `["First", "Second", "Third"]`)
    - **Grading**: Kendall tau distance (counts pairwise inversions)
    - **Normalized Score**: 0.0 to 1.0, based on 1.0 - (inversions / maxInversions)
    - **Partial Credit**: Yes
    
    ### Drag-and-Drop
...
    - **Grading**: Correct placements / total items
...
    
...
    
...
paths:
...
  /progress/graph:
    get:
      tags:
      - Users
      summary: Fetch user progress data for graph.
      operationId: getGraphData
      description: Retrieves user performance data for a radar chart visualization using the KnightWise analytics engine. Returns per-topic weighted performance metrics.
      security:
        - BearerAuth: []
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ProgressGraphResponse'
        401:
          description: Unauthorized
        500:
          description: Server Error
  /progress/history:
    get:
      tags:
      - Users
      summary: Fetch user problem history.
      operationId: getHistory
      description: Retrieves the history of problems solved by the user in a paginated format.
      security:
        - BearerAuth: []
      parameters:
      - name: page
        in: query
        required: false
        type: integer
        description: Page number (default 1, 10 results per page)
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        500:
          description: Server Error
  /progress/messageData:
    get:
      tags:
      - Users
      summary: Fetch user progress message data.
      operationId: getMessage
      description: Returns mastery levels, strongest and weakest topics, and streak data using the KnightWise analytics engine.
      security:
        - BearerAuth: []
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/MessageDataResponse'
        401:
          description: Unauthorized
        500:
          description: Server Error
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Include the token in the Authorization header as 'Bearer your-token'"
definitions:
...
  ProgressGraphResponse:
    type: object
    properties:
      progress:
        type: object
        description: Map of topic name to performance metric data. Keys are Response.TOPIC strings.
        additionalProperties:
          type: object
          properties:
            metric:
              type: number
              example: 0.7342
              description: Point-weighted performance metric from 0.0 to 1.0. Higher is better.
            responseCount:
              type: integer
              example: 12
              description: Number of responses contributing to this topic's metric.
  MessageDataResponse:
    type: object
    properties:
      history:
        type: array
        items:
          type: object
          properties:
            datetime:
              type: string
              example: "2026-03-15T14:30:00.000Z"
            topic:
              type: string
              example: "Arrays"
      mastery:
        type: object
        description: Map of topic name to weighted performance metric (0.0 to 1.0).
        additionalProperties:
          type: number
          example: 0.6821
      strongestTopics:
        type: array
        items:
          type: string
        example: ["Trees", "Recursion", "Sorting"]
        description: Up to 3 topics with the highest performance metrics.
      weakestTopics:
        type: array
        items:
          type: string
        example: ["Variables", "InputOutput", "Branching"]
        description: Up to 3 topics with the lowest performance metrics.
      streak:
        type: integer
        example: 5
        description: Number of consecutive days the user has submitted at least one answer.
host: www.knightwise.dev
basePath: /api
schemes:
...